darkfid::task

Module consensus

Structs

ConsensusInitTaskConfig

Auxiliary structure representing node consensus init task configuration

Functions

consensus_init_task

Sync the node consensus state and start the corresponding task, based on node type.

consensus_task 🔒

Async task used for listening for new blocks and perform consensus.

monitor_network 🔒

Async task to monitor network disconnections.

replicator_task 🔒

Async task to start the consensus task, while monitoring for a network disconnections.