Module darkfid::task::miner

Functions

generate_next_block 🔒

Auxiliary function to generate next block in an atomic manner

generate_transaction 🔒

Auxiliary function to generate a Money::PoWReward transaction

listen_to_network 🔒

Async task to listen for incoming proposals and check if the best fork has changed

mine 🔒

Async task to generate and mine provided fork index next block, while listening for a stop signal

mine_next_block 🔒

Async task to generate and mine provided fork index next block

miner_task

Async task used for participating in the PoW block production. Miner initializes their setup and waits for next finalization, by listenning for new proposals from the network, for optimal conditions. After finalization occurs, they start the actual miner loop, where they first grab the best ranking fork to extend, and start mining procedure for its next block. Additionally, they listen to the network for new proposals, and check if these proposals produce a new best ranking fork. If they do, the stop mining. These two tasks run in parallel, and after one of them finishes, node triggers finallization check.

wait_stop_signal 🔒

Async task to wait for listener’s stop signal.