Module darkfid::proto::protocol_sync

Validator blockchain sync protocol

Structs

ForkSyncRequest

Structure represening a request to ask a node a fork sequence. If we include a specific fork tip, they have to return its sequence, otherwise they respond with their best fork sequence. We also include our own canonical(finalized) tip, so they can verify we follow the same sequence.

ForkSyncResponse

Structure representing the response to ForkSyncRequest, containing the requested fork sequence.

HeaderSyncRequest

Structure represening a request to ask a node for up to BATCH headers before the provided header height.

HeaderSyncResponse

Structure representing the response to HeaderSyncRequest, containing up to BATCH headers before the requested block height.

ProtocolSyncHandler

Handler managing all ProtocolSync messages, over generic P2P protocols.

SyncRequest

Structure represening a request to ask a node for up toBATCH blocks of provided headers.

SyncResponse

Structure representing the response to SyncRequest, containing up to BATCH blocks after the requested block height.

TipRequest

Structure represening a request to ask a node for their current canonical(finalized) tip block hash, if they are synced. We also include our own tip, so they can verify we follow the same sequence.

TipResponse

Structure representing the response to TipRequest, containing a boolean flag to indicate if we are synced, and our canonical(finalized) tip block height and hash.

Constants

BATCH

Functions

handle_receive_fork_request 🔒

Background handler function for ProtocolSyncFork.

handle_receive_header_request 🔒

Background handler function for ProtocolSyncHeader.

handle_receive_request 🔒

Background handler function for ProtocolSync.

handle_receive_tip_request 🔒

Background handler function for ProtocolSyncTip.

Type Aliases

ProtocolSyncHandlerPtr

Atomic pointer to the ProtocolSync handler.