darkfid::task

Module miner

Structs

MinerRewardsRecipientConfig

Auxiliary structure representing node miner rewards recipient configuration

Functions

generate_next_block 🔒

Auxiliary function to generate next block in an atomic manner.

generate_transaction 🔒

Auxiliary function to generate a Money::PoWReward transaction.

listen_to_network 🔒

Async task to listen for incoming proposals and check if the best fork has changed.

mine 🔒

Async task to generate and mine provided fork index next block, while listening for a stop signal.

mine_next_block 🔒

Async task to generate and mine provided fork index next block.

miner_task

Async task used for participating in the PoW block production.

wait_stop_signal

Async task to wait for listener’s stop signal.