darkfid::task

Module sync

Functions

most_common_tip 🔒

Auxiliary function to ask all peers for their current tip and find the most common one.

retrieve_blocks 🔒

Auxiliary function to retrieve blocks of provided headers and apply them to canonical.

retrieve_headers 🔒

Auxiliary function to retrieve headers backwards until our last known one and verify them.

sync_best_fork 🔒

Auxiliary function to retrieve best fork state from a random peer.

sync_task

async task used for block syncing. A checkpoint can be provided to ensure node syncs the correct sequence.

synced_peers 🔒

Auxiliary function to block until node is connected to at least one synced peer, and retrieve the synced peers tips.